requirements.txt
accelerate==1.6.0
celery==5.5.1
datasets==3.5.0
dotenv==0.9.9
Faker==37.1.0
Flask==3.0.3
joblib==1.4.2
opencv-python==4.11.0.86
pdf2image==1.17.0
pypdf==5.4.0
pytesseract==0.3.13
pytest==8.3.3
pytest-mock==3.14.0
python-docx==1.1.2
redis==5.2.1
scikit-learn==1.6.1
torch==2.7.0
transformers==4.51.3